Split relation DSL spike into separate files
