+++ lib/rom-relation.rb
@@ -34,7 +34,11 @@ end # module ROM
 require 'rom/repository'
 require 'rom/environment'
 require 'rom/relation'
+
 require 'rom/schema'
+require 'rom/schema/definition'
+require 'rom/schema/definition/relation'
+require 'rom/schema/definition/relation/base'
 
 require 'rom/header'
 require 'rom/header/attribute_index'
